XMU-Rollcall-Bot-CLI/requirements.txt
pycryptodome
requests
aiohttp
